Split repositories feature support logic
